move everything into class member
